requirements.txt
sanic==22.12.0
SQLAlchemy==1.4.52
asyncpg==0.30.0
alembic==1.11.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.11.9
